requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

langgraph==0.2.38
langchain-core==0.3.22

qdrant-client==1.11.3
sentence-transformers==3.0.1
numpy==1.26.4

httpx==0.27.2
tenacity==8.5.0
